Remove use of AutoLoader::loadClass()
That function is to be removed in Ia6482cce. Regardless, the intent
likely was to check if the BetaFeatures class exists rather than to
simply load the class.

Though in general the current, preferred way to detect an extension
is to use ExtensionRegistry::isLoaded(), class_exists() seems to be
the established way to detect BetaFeatures.

Change-Id: If9ff5b8a2a98285ef9d445483cb0d82f31f2dbf3